Switch from direct Docker builds to Buildkite.com
This adds the capability of using multiarchitecture and multi-OS
agents, all coordinated by the buildkite.com service.

This commit also adds in a new `opam-ci` plugin that can be
used to query the state of the builds from the CLI, as well
as installs an `obi` OCaml library that can be used to build
alternative clients.

Signed-off-by: Anil Madhavapeddy <anil@recoil.org>